Fix various integer conversion problems

+++ kernel/include/kernel/mm/kmalloc.h
@@ -2,8 +2,9 @@
 #define _KERNEL_KMALLOC_H
 
 #include <stddef.h>
+#include <stdint.h>
 
-void kmalloc_init(unsigned lowest_pow2_size, unsigned max_pow2_size);
+void kmalloc_init(uint8_t lowest_pow2_size, uint8_t max_pow2_size);
 
 void *kmalloc(size_t size);
